djangoapp1/requirements.txt
asgiref==3.3.1
attrs==21.2.0
backports.entry-points-selectable==1.1.0
bcrypt==3.2.0
beautifulsoup4==4.9.3
boto3==1.16.52
botocore==1.19.52
cachetools==4.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
cryptography==3.4.8
distlib==0.3.2
distro==1.6.0
dj-database-url==0.5.0
Django==3.1.5
django-audiotracks==0.2.4
django-autoslug==1.9.8
django-bootstrap4==2.3.1
django-cleanup==5.2.0
django-cors-headers==3.8.0
django-crispy-forms==1.10.0
django-embed-video==1.4.0
django-environ==0.4.5
django-filter==2.4.0
django-heroku==0.3.1
django-on-heroku==1.1.2
django-request==1.5.6
django-storages==1.11.1
djangorestframework==3.12.2
docker==5.0.2
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
filelock==3.0.12
google-api-core==1.24.1
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-cloud-core==1.5.0
google-crc32c==1.1.0
google-resumable-media==1.2.0
googleapis-common-protos==1.52.0
gTTS==2.2.1
gTTS-token==1.1.4
gunicorn==20.1.0
idna==2.10
imagesize==1.2.0
jmespath==0.10.0
jsonschema==3.2.0
mutagen==1.45.1
mutagenx==1.22
oauthlib==3.1.0
paramiko==2.7.2
Pillow==8.1.0
pipenv==2021.5.29
platformdirs==2.3.0
protobuf==3.14.0
psycopg2-binary==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydub==0.24.1
PyNaCl==1.4.0
pyrsistent==0.18.0
python-dateutil==2.8.1
python-dotenv==0.19.0
pytz==2020.5
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7
s3transfer==0.3.4
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
texttable==1.6.4
typing-extensions==3.10.0.0
urllib3==1.26.2
virtualenv==20.7.2
virtualenv-clone==0.5.6
websocket-client==0.59.0
whitenoise==5.2.0
